+++ packages/google_maps_flutter/google_maps_flutter_ios_sdk9/lib/src/google_maps_flutter_ios.dart
@@ -1162,6 +1162,9 @@ PlatformMapConfiguration _platformMapConfigurationFromMapConfiguration(MapConfig
     markerType: _platformMarkerTypeFromMarkerType(config.markerType ?? MarkerType.marker),
     mapId: config.mapId,
     style: config.style,
+    backgroundColor: config.backgroundColor == null
+        ? null
+        : _platformColorFromColor(config.backgroundColor!),
   );
 }
 
@@ -1219,6 +1222,9 @@ PlatformMapConfiguration _platformMapConfigurationFromOptionsJson(Map<String, Ob
     markerType: PlatformMarkerType.marker,
     mapId: options['mapId'] as String?,
     style: options['style'] as String?,
+    backgroundColor: options['backgroundColor'] == null
+        ? null
+        : _platformColorFromColor(Color(options['backgroundColor']! as int)),
   );
 }